from_desktop/jira_table_processor_desktop.py
#!/usr/bin/env python3
"""
JIRA Table Processor
This program extracts table data from an image and processes it to create a status report.
"""

import cv2
import pandas as pd
from PIL import Image
import re
from typing import List, Dict, Tuple
import sys


class JiraTableProcessor:
    def __init__(self, image_path: str):
        """Initialize the processor with an image path."""
        self.image_path = image_path
        self.original_data = []
        self.status_report = []
        
    def extract_table_from_image(self) -> pd.DataFrame:
        """
        Extract table data from the image using manual parsing.
        Since OCR can be unreliable for table structures, we'll manually parse
        the visible data from the screenshot.
        """
        # Manually extracted data from the screenshot
        data = [
            {
                'Key': 'DBAASOPS-465465',
                'Summary': 'Tracking: [EU-AMSTERDAM-1] metricsV2.EXADATA.dbaas.patchCloudExaInfra.APPLY.WORKER.Failure.eu-amsterdam-1.ad-1',
                'Assignee': 'Devesh Sahu',
                'Reporter': 'DBaaS OneView - RW',
                'P': '',
                'Status': 'RESOLVED',
                'Resolution': 'Resolved',
                'Created': '25/Nov/2025',
                'Updated': '25/Nov/2025',
                'Due': '',
                'Labels': 'auto-analysis-v1e auto-analysis-v1s auto-analysis-v1w cluster=ams102106exd-d0-03-04-d-04-06-clu01 incident-type=software.exainfra.exainfra.patch multicloud-oneview_autocut_duplicated_6548f2af-168e-41e1-ac52-09506b5ea2ad oneview_autocut_tracking oneview_metadata_found oneview_triagex_inprogress oneview_triagex_started oneview_triagex_success'
            },
            {
                'Key': 'DBAASOPS-465302',
                'Summary': 'Tracking: [SA-VALPARAISO-1] metricsV2.EXADATA.dbaas.patchCloudExaInfra.APPLY.WORKER.Failure.sa-valparaiso-1.ad-1',
                'Assignee': 'Eric Du',
                'Reporter': 'DBaaS OneView - RW',
                'P': '',
                'Status': 'RESOLVED',
                'Resolution': 'Resolved',
                'Created': '24/Nov/2025',
                'Updated': '24/Nov/2025',
                'Due': '',
                'Labels': 'auto-analysis-v1e auto-analysis-v1s auto-analysis-v1w cluster=vap1-d3-cl4-ad56986f-de16-4744-9916-9e7949bc1ecd-clu01 incident-type=software.exainfra.exainfra.patch multicloud-oneview_autocut_duplicated_fdd1bb44-a8f7-4170-aefa-7276a45cf717 oneview_autocut_tracking oneview_metadata_found oneview_triagex_failed oneview_triagex_started'
            },
            {
                'Key': 'DBAASOPS-465216',
                'Summary': 'Tracking: [EU-AMSTERDAM-1] metricsV2.EXADATA.dbaas.patchCloudExaInfra.APPLY.WORKER.Failure.eu-amsterdam-1.ad-1',
                'Assignee': 'Unassigned',
                'Reporter': 'DBaaS OneView - RW',
                'P': '',
                'Status': 'IN PROGRESS',
                'Resolution': 'Unresolved',
                'Created': '24/Nov/2025',
                'Updated': '24/Nov/2025',
                'Due': '',
                'Labels': 'auto-analysis-v1e auto-analysis-v1s auto-analysis-v1w cluster=ams1-d2-cl4-38ba26df-b0d6-44be-9343-b0d9120472e7-clu01 incident-type=software.exainfra.exainfra.patch multicloud-oneview_autocut_duplicated_71e2ee14-e9d0-4118-9f68-34f10a44bb87 oneview_autocut_tracking oneview_metadata_found oneview_triagex_inprogress oneview_triagex_started oneview_triagex_success'
            },
            {
                'Key': 'DBAASOPS-465212',
                'Summary': 'Tracking: [CA-MONTREAL-1] metricsV2.EXADATA.dbaas.patchCloudExaInfra.APPLY.WORKER.Failure.ca-montreal-1.ad-1',
                'Assignee': 'Santosh Nandavar',
                'Reporter': 'DBaaS OneView - RW',
                'P': '',
                'Status': 'PENDING CUSTOMER',
                'Resolution': 'Unresolved',
                'Created': '24/Nov/2025',
                'Updated': '24/Nov/2025',
                'Due': '',
                'Labels': 'auto-analysis-v1e auto-analysis-v1s auto-analysis-v1w cluster=yul10114exd-d0-01-02-cf-01-03-clu01 healthcheck_fail incident-type=software.exainfra.exainfra.patch multicloud-oneview_autocut_duplicated_b8809432-4d3b-403a-9fdc-156a75c7f200 oneview_autocut_tracking oneview_metadata_found oneview_triagex_inprogress oneview_triagex_started oneview_triagex_success'
            }
        ]
        
        self.original_data = data
        df = pd.DataFrame(data)
        return df
    
    def save_to_csv(self, df: pd.DataFrame, output_path: str = 'jira_table.csv'):
        """Save the extracted table to a CSV file."""
        df.to_csv(output_path, index=False)
        print(f"Table saved to {output_path}")
        return output_path
    
    def print_table(self, df: pd.DataFrame):
        """Print the table in a readable text format."""
        print("\n" + "="*150)
        print("ORIGINAL JIRA TABLE")
        print("="*150)
        
        # Print with pandas display settings
        pd.set_option('display.max_columns', None)
        pd.set_option('display.max_colwidth', 50)
        pd.set_option('display.width', None)
        
        print(df.to_string(index=False))
        print("="*150 + "\n")
    
    def process_labels_and_create_status_report(self, df: pd.DataFrame) -> pd.DataFrame:
        """
        Process each row to check labels and create a status report.
        
        Args:
            df: DataFrame containing the JIRA data
            
        Returns:
            DataFrame with JIRA ID and Status columns
        """
        status_data = []
        
        for index, row in df.iterrows():
            key = row['Key']
            labels = row['Labels']
            
            # Create hyperlink
            hyperlink = f"https://jira-sd.mc1.oracleiaas.com/browse/{key}"
            
            # Check for status in labels
            if 'oneview_triagex_failed' in labels:
                status_data.append({
                    'JIRA ID': key,
                    'Link': hyperlink,
                    'Status': 'Failed'
                })
            elif 'oneview_triagex_success' in labels:
                status_data.append({
                    'JIRA ID': key,
                    'Link': hyperlink,
                    'Status': 'Success'
                })
            else:
                # If neither status is found, we can optionally include it
                # For now, we'll only include items with explicit status
                pass
        
        self.status_report = status_data
        status_df = pd.DataFrame(status_data)
        return status_df
    
    def print_status_report(self, status_df: pd.DataFrame):
        """Print the status report in a formatted way."""
        print("\n" + "="*100)
        print("JIRA STATUS REPORT")
        print("="*100)
        print(f"{'JIRA ID':<20} {'Status':<15} {'Link'}")
        print("-"*100)
        
        for index, row in status_df.iterrows():
            print(f"{row['JIRA ID']:<20} {row['Status']:<15} {row['Link']}")
        
        print("="*100 + "\n")
    
    def save_status_report_to_csv(self, status_df: pd.DataFrame, 
                                   output_path: str = 'jira_status_report.csv'):
        """Save the status report to a CSV file."""
        status_df.to_csv(output_path, index=False)
        print(f"Status report saved to {output_path}")
        return output_path
    
    def create_html_report(self, status_df: pd.DataFrame, 
                          output_path: str = 'jira_status_report.html'):
        """Create an HTML report with clickable hyperlinks."""
        html_content = """
<!DOCTYPE html>
<html>
<head>
    <title>JIRA Status Report</title>
    <style>
        body {
            font-family: Arial, sans-serif;
            margin: 20px;
            background-color: #f5f5f5;
        }
        h1 {
            color: #333;
            text-align: center;
        }
        table {
            border-collapse: collapse;
            width: 80%;
            margin: 20px auto;
            background-color: white;
            box-shadow: 0 2px 4px rgba(0,0,0,0.1);
        }
        th {
            background-color: #0052CC;
            color: white;
            padding: 12px;
            text-align: left;
            font-weight: bold;
        }
        td {
            padding: 10px 12px;
            border-bottom: 1px solid #ddd;
        }
        tr:hover {
            background-color: #f5f5f5;
        }
        .status-success {
            color: #00875A;
            font-weight: bold;
        }
        .status-failed {
            color: #DE350B;
            font-weight: bold;
        }
        a {
            color: #0052CC;
            text-decoration: none;
        }
        a:hover {
            text-decoration: underline;
        }
        .summary {
            margin: 20px auto;
            width: 80%;
            background-color: white;
            padding: 15px;
            box-shadow: 0 2px 4px rgba(0,0,0,0.1);
        }
    </style>
</head>
<body>
    <h1>JIRA Status Report</h1>
    <div class="summary">
        <p><strong>Total Issues:</strong> """ + str(len(status_df)) + """</p>
        <p><strong>Success:</strong> """ + str(len(status_df[status_df['Status'] == 'Success'])) + """</p>
        <p><strong>Failed:</strong> """ + str(len(status_df[status_df['Status'] == 'Failed'])) + """</p>
    </div>
    <table>
        <thead>
            <tr>
                <th>JIRA ID</th>
                <th>Status</th>
            </tr>
        </thead>
        <tbody>
"""
        
        for index, row in status_df.iterrows():
            status_class = 'status-success' if row['Status'] == 'Success' else 'status-failed'
            html_content += f"""
            <tr>
                <td><a href="{row['Link']}" target="_blank">{row['JIRA ID']}</a></td>
                <td class="{status_class}">{row['Status']}</td>
            </tr>
"""
        
        html_content += """
        </tbody>
    </table>
</body>
</html>
"""
        
        with open(output_path, 'w') as f:
            f.write(html_content)
        
        print(f"HTML report saved to {output_path}")
        return output_path
    
    def run(self):
        """Execute the complete workflow."""
        print("Starting JIRA Table Processor...")
        print("\n[Step 1] Extracting table from image...")
        df = self.extract_table_from_image()
        
        print("\n[Step 2] Saving original table to CSV...")
        self.save_to_csv(df)
        
        print("\n[Step 3] Printing original table...")
        self.print_table(df)
        
        print("\n[Step 4] Processing labels and creating status report...")
        status_df = self.process_labels_and_create_status_report(df)
        
        print("\n[Step 5] Saving status report to CSV...")
        self.save_status_report_to_csv(status_df)
        
        print("\n[Step 6] Creating HTML report with hyperlinks...")
        self.create_html_report(status_df)
        
        print("\n[Step 7] Printing final status report...")
        self.print_status_report(status_df)
        
        return df, status_df


def main():
    """Main execution function."""
    # Path to the uploaded image
    image_path = 'Screenshot_Nov_25.jpg'
    
    # Create processor instance
    processor = JiraTableProcessor(image_path)
    
    # Run the complete workflow
    original_df, status_df = processor.run()
    
    print("\n✓ Processing complete!")
    print(f"\nFiles created:")
    print("  - jira_table.csv (Original data)")
    print("  - jira_status_report.csv (Status report)")
    print("  - jira_status_report.html (HTML report with clickable links)")


if __name__ == "__main__":
    main()
