from_desktop/requirements.txt
# Requirements for JIRA Table Processor
# Python 3.8+

# Core data processing
pandas>=2.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Optional: For enhanced table extraction (if needed in future)
# pytesseract>=0.3.10

# Optional: For Excel export support
openpyxl>=3.1.0
